+++ python/cugraph/cugraph/gnn/dgl_extensions/utils/__init__.py
@@ -10,6 +10,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from .graph_store import CuGraphStore
-from .dgl_extensions.feature_storage import CuFeatureStorage
